OPENGL: Move pipeline code to pipelines/.
